Changed witnessIdComponent to being reusable for styling and click function binding. Now witness labels are shown in the witnessModal component too.
