+++ src/app/components/witness-metadata/witness-metadata.component.html
@@ -1,5 +1,5 @@
 <ng-container *ngFor="let witId of this.witnessIds; index as i">
     <span *ngIf="i === 0">{{ ' ' }}</span>
-    <evt-witness-id [witnessId]="witId"></evt-witness-id>
+    <evt-witness-id style="font-size: 90%;" [witnessId]="witId" (witnessClicked)="onWitnessClicked($event)"></evt-witness-id>
     <span *ngIf="i !== (this.witnessIds.length - 1)">{{ ' ' }}</span>
 </ng-container>
