+++ src/app/view-modes/collation/modal-witness-item/modal-witness-item.component.html
@@ -2,7 +2,8 @@
     <div class="modal-item d-flex justify-content-between align-items-center py-1" (click)="toggleCollapse()">
         <evt-icon role="button" *ngIf="witness.witnesses.length" [iconInfo]="chevronIcon"
             [attr.aria-expanded]="!isCollapsed"></evt-icon>
-        <span class="ms-2 fs-5">{{ witness.label }}</span>
+        <!-- <span class="ms-2 fs-5">{{ witness.label }}</span> -->
+        <evt-witness-id class="ms-2 fs-5" [witnessId]="witness.id"></evt-witness-id>
 
         <evt-button *ngIf="witness.canSelect" type="button" class="ms-auto" additionalClasses="btn-sm btn-floating"
             [label]="'add' | translate" (btnClick)="onSelectClicked(this.witness.id)"></evt-button>
